% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aldous_broder}
\alias{aldous_broder}
\title{Sample a random spanning tree}
\usage{
aldous_broder(W, start)
}
\arguments{
\item{W}{a matrix.}

\item{start}{an non-negative integer indicating the starting point.}
}
\value{
An adjacency matrix of the random spanning tree.
}
\description{
Sample a random spanning tree
}

