% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_cover}
\alias{fast_cover}
\title{Sample a random spanning tree}
\usage{
fast_cover(W, start, threshold)
}
\arguments{
\item{W}{a matrix.}

\item{start}{an non-negative integer indicating the starting point.}

\item{threshold}{an non-negative integer indicating the threshold for deploying fast-forwarding.}
}
\value{
An adjacency matrix of the random spanning tree.
}
\description{
Sample a random spanning tree
}
