% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{plot_as_tree}
\alias{plot_as_tree}
\title{plot a tree graph with root specified}
\usage{
plot_as_tree(tree, root = "1")
}
\arguments{
\item{root}{the vertex that serves as the root}

\item{g}{A graph object from igraph}
}
\value{
A plot of the tree.
}
\description{
plot a tree graph with root specified
}
