% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{plot_underlying}
\alias{plot_underlying}
\title{plot the underlying graph}
\usage{
plot_underlying(g)
}
\arguments{
\item{g}{A graph object from igraph}
}
\value{
A plot of the graph.
}
\description{
plot the underlying graph
}
