% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_helpers.R
\name{plot_as_tree}
\alias{plot_as_tree}
\title{plot a tree graph with root specified}
\usage{
plot_as_tree(tree, root = "1", vertex_size = 50)
}
\arguments{
\item{tree}{A graph object from igraph}

\item{root}{the vertex that serves as the root}

\item{vertex_size}{The size of the vertices in the plot}
}
\value{
A plot of the tree.
}
\description{
plot a tree graph with root specified
}
